[front] refactor: replace several MUI <Link> by <ExternalLink>
