+++ frontend/src/features/comparisons/ComparisonHelperPresidentielle2022.tsx
@@ -1,6 +1,8 @@
 import React, { useState, useEffect } from 'react';
 
-import { Typography, Link, Box, Grid } from '@mui/material';
+import { Typography, Box, Grid } from '@mui/material';
+
+import { ExternalLink } from 'src/components';
 import { getAllCandidates } from 'src/utils/polls/presidentielle2022';
 import { EntityObject } from 'src/utils/types';
 
@@ -44,14 +46,13 @@ const ComparisonHelperPresidentielle2022 = () => {
                 ({ metadata }) =>
                   metadata && (
                     <li key={metadata.name}>
-                      <Link
-                        color="text.secondary"
+                      <ExternalLink
                         href={metadata.website_url}
-                        rel="noopener"
                         target="_blank"
+                        sx={{ color: 'text.secondary' }}
                       >
                         {metadata.name}
-                      </Link>
+                      </ExternalLink>
                     </li>
                   )
               )}
@@ -61,57 +62,52 @@ const ComparisonHelperPresidentielle2022 = () => {
             <Typography variant="h6">📰 Comparateurs de programmes</Typography>
             <ul>
               <li>
-                <Link
-                  color="text.secondary"
+                <ExternalLink
                   href="https://www.lemonde.fr/les-decodeurs/article/2022/02/16/election-presidentielle-2022-comparez-les-programmes-des-principaux-candidats_6113964_4355770.html"
-                  rel="noopener"
                   target="_blank"
+                  sx={{ color: 'text.secondary' }}
                 >
                   Du journal Le Monde
-                </Link>
+                </ExternalLink>
               </li>
               <li>
-                <Link
-                  color="text.secondary"
+                <ExternalLink
                   href="https://www.humanite.fr/comparateur-des-programmes-presidentielle-2022"
-                  rel="noopener"
                   target="_blank"
+                  sx={{ color: 'text.secondary' }}
                 >
                   Du journal l&apos;Humanité
-                </Link>
+                </ExternalLink>
               </li>
               <li>
-                <Link
-                  color="text.secondary"
+                <ExternalLink
                   href="https://comparateur-programmes.lefigaro.fr/"
-                  rel="noopener"
                   target="_blank"
+                  sx={{ color: 'text.secondary' }}
                 >
                   Du journal Le Figaro
-                </Link>
+                </ExternalLink>
               </li>
             </ul>
             <Typography variant="h6">🌳 Analyses des mesures climat</Typography>
             <ul>
               <li>
-                <Link
-                  color="text.secondary"
+                <ExternalLink
                   href="https://reseauactionclimat.org/presidentielle-candidats-climat/"
-                  rel="noopener"
                   target="_blank"
+                  sx={{ color: 'text.secondary' }}
                 >
                   par le Réseau Action Climat France
-                </Link>
+                </ExternalLink>
               </li>
               <li>
-                <Link
-                  color="text.secondary"
+                <ExternalLink
                   href="https://presidentielle2022.theshifters.org/decryptage/"
-                  rel="noopener"
                   target="_blank"
+                  sx={{ color: 'text.secondary' }}
                 >
                   par The Shifters
-                </Link>
+                </ExternalLink>
               </li>
             </ul>
           </Grid>
@@ -125,16 +121,15 @@ const ComparisonHelperPresidentielle2022 = () => {
             ({ metadata }) =>
               metadata && (
                 <li key={metadata.name}>
-                  <Link
-                    color="text.secondary"
+                  <ExternalLink
                     href={`https://fr.wikipedia.org/wiki/${encodeURIComponent(
                       metadata.frwiki_title
                     )}`}
-                    rel="noopener"
                     target="_blank"
+                    sx={{ color: 'text.secondary' }}
                   >
                     {metadata.name}
-                  </Link>
+                  </ExternalLink>
                 </li>
               )
           )}
